+++ src/main/resources/scheduler/assets/js/services.js
@@ -167,10 +167,13 @@
 
           getJobUpdateDetails: function (updateKey) {
             return async(function (deferred) {
-              auroraClient.getSchedulerClient().getJobUpdateDetails(updateKey, function (response) {
+              var schedulerClient = auroraClient.getSchedulerClient();
+              var query = new JobUpdateQuery();
+              query.key = updateKey;
+              schedulerClient.getJobUpdateDetails(null, query, function (response) {
                 var result = auroraClient.processResponse(response);
                 result.details = response.result !== null ?
-                  response.result.getJobUpdateDetailsResult.details : {};
+                  response.result.getJobUpdateDetailsResult.detailsList[0] : {};
                 deferred.resolve(result);
               });
             });